Improve the device list and the connection code
The device list allows us to support multiple device IDs (though not
simultaneously yet) with different capabilities. The connection code
should also produce more meaningful error messages now.

The different capabilities are handled via attributes in the device
list. These attributes allow specifying the maximum mouse dpi, as well
as the presence or absence of a battery.

The DPI setting needs to be investigated: do mice with different battery
settings report a different value, or does it affect the granularity
and/or the scale factor used for the raw value calculation?